Add GitRef() function to more robustly extract the commit SHA
We used to rely on the file .git/ref, added by the Concourse git resource
to a repository. This breaks because it behaves as follows:
- if the git resource doesn't have a `tag_filter`, then it contains the SHA
- else, it contains the tag name
while we always want the SHA, to then pass to the GitHub API.

Now we look directly into the .git directory and follow the trail from
.git/HEAD to .git/refs/heads/<branch-name>, which is guaranteed to contain
the SHA of the commit at HEAD.

This required also to change the CopyDir() helper (see previous commit) and
to change the testdata layout, which now is:

    resource/testdata/a-repo/
    └── dot.git
        ├── HEAD.template
        ├── config.template
        └── refs
            └── heads
                └── {{.branch_name}}.template

See the added test TestGitRef for how everything works together.