Ensure we always return a non-null version.
We were already always returning a non-null version for the put and
check steps. This commit does the same also for the get step.

Why? Returning a null version was crashing the ATC for a put step (see
https://github.com/concourse/concourse/pull/4442, fixed in Concourse 5.6.0).
We err on the side of caution and ensure (and explicitly tests) we do the
same also fot check and get.