handle http server error code response
`StatusInternalServerError` response is now handled and
text error code,  server originator, are added in response.
Previously, only the code was reported with an empty response body
because the handling was left to the `default` case.
As a bonus, `StatusNotFound` will report the text code as well.
Fix parameter `access_token` in README

The tests cover now all the 3 cases (success, not found, Internal Server Error).