provide debug info as soon as possible, defer input validation
This changes the logic of previous PR #85 and makes the code a bit more complicated,
with the goal of providing more debug information if something goes wrong at input
validation time.

This chicken-and-egg problem is due to the fact that logging configuration is passed
too late, at the same time as all the other resource Source configuration, so to give
as much debugging information as possible we need to get the log level as soon as
possible, also if the Source has other errors. This cannot be simplified, we are
working within the limits of the Concourse resource protocol.

PCI-2587