github API: give hints in case of status code 401 (unauthorized)
In a pursuit of security that I do not understand, if a Personal Access Token (PAT) is expired,
the GitHub API returns the same HTTP status code (401) as if the credentials were wrong.

This happened to me yesterday, it was quite confusing and made me waste time to troubleshoot.

Since there is no API endpoint to validate whether a token is valid or expired, we do the
best we can: we give a hint to the user to check his/her email :-/

PCI-2587