googlechat: fix bug in encoding the thread key as URL parameter
Before, we were appending the URL parameters to the webhook URL and then pass such URL to
http.NewRequestWithContext(). This worked because the webhook URL did not contain a port.

In this PR, I added tests with the SpyHttpServer, which (obviously) returns a URL with a port,
http.NewRequestWithContext() parses the URL, sees the port and is tripped by the appended
URL parameter... so arguably it is http.NewRequestWithContext() that does not validate the URL
in a uniform way.

Now, we set the URL parameters after the call to http.NewRequestWithContext(). This should be
the correct way.

PCI-2587