review: simplify log_level backward compatibility
if somebody was using "silent" (but why??? in any case the resource logs are
not visible in the build output...) now he/she will get "info" instead. If
he notices, he will simply set to "off".

Thinking more about it, I think that the original API design error was to
allow to set the log_level, and especially to set the log_level with the syslog
classic granularity, which is useless.

Instead, the source configuration should have simply offered a boolean "debug",
with default false. To keep in mind for the next resource we write...