the GitHub repo is now an option
Cogito is able now sending to chat only

method Sinks() now parse resource parameters before to return the configured sinks

If put.params.sinks is left unconfigured, it behaviors as before.
If put.params.sinks is configured, it returns the wanted sinks.

If the sink is mispelled and consequently unknown it will throw an error.