add sinks parameter in source configuration
The `sinks` parameter rapresents a list of supported sinks (currently only github and gchat).
Its content, informs Cogito about selective sending of messages/statuses.

In the case of chat only sending, if there is no message file,
the inputs dir is not mandatory anymore. It's still mandatory
if there is a message file in the put parameters.