github/commitstatus: minimal fix for zero sleep time
We were using the value 0 for time to sleep as a sentinel, meaning to
stop attempting. But then we realized that there were legitimate
(although corner cases) for a value of 0, and in those cases we actually
meant to attempt once more.

The fix is to be more explicit: add a boolean to indicate whether
we should retry or not, instead of treating 0 as a sentinel.

All tests are passing.